Put stuff in their correct folders
